backend/src/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
filelock==3.0.12
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.2
whitenoise==5.2.0
